fix: KanbanTestPage 삭제, Kanban Component로 수정
#33
